fix: notice, info 폴더명 변경
notice 페이지에 info가, info 페이지에 notice가 들어있었음
