+++ pages/store/[storeId]/info/info.stories.tsx
