+++ pages/store/[storeId]/notice/notice.styled.ts
